Window row/col can be fractional

+++ crates/nvi/src/nvim/types.rs
@@ -570,9 +570,9 @@ pub struct WindowConf {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bufpos: Option<(u64, u64)>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub row: Option<i64>,
+    pub row: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub col: Option<i64>,
+    pub col: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub focusable: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
